// src/constants/analyticsEvents.enum.ts
export enum AnalyticsEvents {
    PATIENT_ON_BOARD = 'Patient On Board',
    PATIENT_DOCTOR_SEARCH = 'Patient Doctor Search',
    PATIENT_ACITIVITY = 'Patient Activity',

    DOCTOR_ON_BOARD = 'Doctor On Board',
    DOCTOR_EPRESCRIPTION = 'Doctor E-Prescription',
    DOCTOR_APPOINTMENTS = 'Doctor Appointments',

    PHARMACY_ONBOARDING = 'Pharmacy on Board',
    PHARMACY_EPRESCRIPTION = 'e prescription',

    LABORATORY_ONBOARDING = 'laboratory on Board',
    LABORATORY_EPRESCRIPTION = 'lab e prescription',
}